+++ app/javascript/packs/state/actions/user.js
@@ -1,0 +1,11 @@
+import {signOut} from '../../api';
+
+export const actions = {
+  SIGN_OUT: 'SIGN_OUT'
+}
+
+export function signOutUser() {
+  return dispatch => {
+    return signOut().then(d => dispatch({type: actions.SIGN_OUT}));
+  }
+}