+++ app/javascript/packs/state/reducers/user.js
@@ -1,0 +1,13 @@
+import { actions } from '../actions/user'
+
+const user = (state = [], action) => {
+  switch (action.type) {
+    case actions.SIGN_OUT:
+      location.reload();
+      return {...state}
+    default:
+      return state
+  }
+}
+
+export default user